requirements.txt
pandas==1.1.3
uvicorn==0.11.6
fastapi==0.60.1
google-cloud-storage==1.25.0